+++ docs/project/changelog.md
@@ -34,9 +34,18 @@ myst:
   {pr}`5597`
 - {{ Fix }} Add the current working directory to the path instead of `$HOME`.
   {pr}`5630`
+- {{ Fix }} Fixed a fatal error when stack switching is enabled on a function
+  that raises an asynchronous error.
+  {pr}`5678`
+
+- {{ Breaking }} When `lockfileURL` is given to `loadPyodide`, the
+  base URL for the packages is now calculated from the lockfile URL, not from
+  the `indexURL`.
+  {pr}`5652`
 
 - {{ Enhancement }} `pyodide.loadPackage` now prints the output to the `stdout` and `stderr`
   streams that are passed to `loadPyodide()` or by `pyodide.setStdout()` and `pyodide.setStderr()`.
+  {pr}`5621`
 
 ### `python` CLI entrypoint